+++ control.lua
@@ -6,6 +6,8 @@ if dbg.flags.add_debug_commands then
     commands.add_command(dbg.commands.print_off, nil, lib.on_console_command)
     commands.add_command(dbg.commands.print_on, nil, lib.on_console_command)
     commands.add_command(dbg.commands.print_connect_lamps, nil, lib.on_console_command)
+    commands.add_command(dbg.commands.print_on_research_finished, nil, lib.on_console_command)
+    commands.add_command(dbg.commands.print_set_rx_filter, nil, lib.on_console_command)
     commands.add_command(dbg.commands.print_update_receiver_filters, nil, lib.on_console_command)
     commands.add_command(dbg.commands.print_update_net_id, nil, lib.on_console_command)
 end